get my type right

+++ Tests/Helpers/UInt32+Test.swift
@@ -1,4 +1,4 @@
 extension UInt32 {
   public static let testSequence: UInt32 = 1
-  public static let testTimestamp: UInt64 = 0
+  public static let testTimestamp: UInt32 = 0
 }
